move grovestark prover to model
