+++ src/model/mod.rs
@@ -1,5 +1,6 @@
 pub mod amount;
 pub mod contested_name;
+pub mod grovestark_prover;
 pub mod password_info;
 pub mod proof_log_item;
 pub mod qualified_contract;
