--- src/proofs/mod.rs
@@ -1,1 +1,0 @@
-pub mod grovestark_integration;
